+++ packages/sdk/src/utils/transaction.ts
@@ -1,6 +1,6 @@
-import { Address, PublicClient, serializeTransaction } from 'viem'
+import { Address, PublicClient } from 'viem'
 import { LogLevel, getClient } from '..'
-import { Execute, ReservoirWallet, TransactionStepItem, paths } from '../types'
+import { Execute, ReservoirWallet, TransactionStepItem } from '../types'
 import { CrossChainTransactionError, TransactionTimeoutError } from '../errors'
 import axios, {
   AxiosRequestConfig,
@@ -36,15 +36,6 @@ export async function sendTransactionSafely(
   const reservoirChain =
     client.chains.find((chain) => chain.id == chainId) || null
   let txHash = await wallet.handleSendTransactionStep(chainId, item, step)
-  submitTransactionToSolver({
-    chainId,
-    viemClient,
-    step,
-    request,
-    headers,
-    txHash,
-    isCrossChainIntent,
-  })
   const pollingInterval = reservoirChain?.checkPollingInterval ?? 5000
   const maximumAttempts =
     client.maxPollingAttemptsBeforeTimeout ??
@@ -77,15 +68,6 @@ export async function sendTransactionSafely(
           ['Transaction replaced', replacement],
           LogLevel.Verbose
         )
-        submitTransactionToSolver({
-          chainId,
-          viemClient,
-          step,
-          request,
-          headers,
-          txHash,
-          isCrossChainIntent,
-        })
       },
     })
     .catch((error) => {
@@ -191,77 +173,4 @@ export async function sendTransactionSafely(
   }
 
   return true
-}
-
-const submitTransactionToSolver = async ({
-  chainId,
-  viemClient,
-  request,
-  headers,
-  step,
-  txHash,
-  isCrossChainIntent,
-}: {
-  chainId: number
-  viemClient: PublicClient
-  step: Execute['steps'][0]
-  request: AxiosRequestConfig
-  headers: AxiosRequestHeaders
-  txHash: Address | undefined
-  isCrossChainIntent?: boolean
-}) => {
-  if (step.id === 'sale' && txHash && isCrossChainIntent) {
-    getClient()?.log(['Submitting transaction to solver'], LogLevel.Verbose)
-    try {
-      const tx = await viemClient.getTransaction({
-        hash: txHash,
-      })
-      const serializedTx = serializeTransaction(
-        {
-          type: tx.type,
-          chainId: tx.chainId ?? chainId,
-          gas: tx.gas,
-          nonce: tx.nonce,
-          to: tx.to || undefined,
-          value: tx.value,
-          maxFeePerGas: tx.maxFeePerGas,
-          maxPriorityFeePerGas: tx.maxPriorityFeePerGas,
-          accessList: tx.accessList,
-          data: tx.input,
-        },
-        {
-          v: tx.v < 27n ? tx.v + 27n : tx.v,
-          r: tx.r,
-          s: tx.s,
-        }
-      )
-
-      const solveData: paths['/execute/solve/v1']['post']['parameters']['body']['body'] =
-        {
-          //@ts-ignore
-          kind: 'cross-chain-intent',
-          tx: serializedTx,
-          chainId: chainId,
-        }
-
-      axios
-        .request({
-          url: `${request.baseURL}/execute/solve/v1`,
-          method: 'POST',
-          headers: headers,
-          data: solveData,
-        })
-        .then(() => {
-          getClient()?.log(
-            ['Transaction submitted to solver'],
-            LogLevel.Verbose
-          )
-        })
-    } catch (e) {
-      getClient()?.log(
-        ['Failed to submit transaction to solver', e],
-        LogLevel.Warn
-      )
-    }
-  }
 }