+++ packages/ui/src/modal/buy/BuyModalRenderer.tsx
@@ -250,13 +250,7 @@ export const BuyModalRenderer: FC<Props> = ({
       paymentCurrency: EnhancedCurrency | undefined,
       paymentTokens: EnhancedCurrency[]
     ) => {
-      if (
-        !open ||
-        !client ||
-        !tokenData ||
-        !token ||
-        paymentTokens.length === 0
-      ) {
+      if (!open || !client || !tokenData || !token) {
         setPath(undefined)
         return
       }
@@ -295,7 +289,7 @@ export const BuyModalRenderer: FC<Props> = ({
         if (paymentCurrency.chainId) {
           options.currencyChainId = paymentCurrency.chainId
         }
-      } else if (!includeListingCurrency) {
+      } else if (!includeListingCurrency && paymentTokens[0]) {
         options.currency = paymentTokens[0].address
         if (paymentTokens[0].chainId) {
           options.currencyChainId = paymentTokens[0].chainId