+++ packages/gelato-adapter/CHANGELOG.md
@@ -1,5 +1,55 @@
 # @reservoir0x/gelato-adapter
 
+## 1.0.9
+
+### Patch Changes
+
+- Updated dependencies [fd61f2b]
+- Updated dependencies [5a0aa7f]
+  - @reservoir0x/reservoir-sdk@2.0.9
+
+## 1.0.8
+
+### Patch Changes
+
+- Updated dependencies [c2ef89f]
+  - @reservoir0x/reservoir-sdk@2.0.8
+
+## 1.0.7
+
+### Patch Changes
+
+- Updated dependencies [213a434]
+  - @reservoir0x/reservoir-sdk@2.0.7
+
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies [61b2d03]
+  - @reservoir0x/reservoir-sdk@2.0.6
+
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies [9d802e7]
+  - @reservoir0x/reservoir-sdk@2.0.5
+
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies [46c7295]
+  - @reservoir0x/reservoir-sdk@2.0.4
+
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies [4bc7810]
+  - @reservoir0x/reservoir-sdk@2.0.3
+
 ## 1.0.2
 
 ### Patch Changes